kettlebell_app/requirements.txt
asgiref==3.8.1
autoflake==2.3.1
black==24.8.0
channels==4.1.0
click==8.1.7
diff-match-patch==20230430
Django==5.1
django-import-export==4.1.1
flake8==7.1.1
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
pyyaml==6.0.2
gunicorn==23.0.0
